fix(authority-source-files): refactor code
- Rename and replace util class
- Change default constraint message value
- Modify tests

Closes: MODELINKS-210
