+++ src/main/java/org/folio/entlinks/controller/ApiErrorHandler.java
@@ -3,10 +3,10 @@ package org.folio.entlinks.controller;
 import static java.util.Collections.emptyList;
 import static org.apache.logging.log4j.Level.DEBUG;
 import static org.apache.logging.log4j.Level.WARN;
-import static org.folio.entlinks.config.constants.Constrains.getErrorCode;
 import static org.folio.entlinks.config.constants.ErrorCode.NOT_EXISTED_AUTHORITY_SOURCE_FILE;
 import static org.folio.entlinks.exception.type.ErrorType.UNKNOWN_ERROR;
 import static org.folio.entlinks.exception.type.ErrorType.VALIDATION_ERROR;
+import static org.folio.entlinks.utils.DatabaseConstraintTranslator.translate;
 import static org.springframework.http.HttpStatus.BAD_REQUEST;
 import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
 import static org.springframework.http.HttpStatus.NOT_FOUND;
@@ -134,8 +134,7 @@ public class ApiErrorHandler {
   public ResponseEntity<Errors> conflict(Exception e) {
     var cause = e.getCause();
     if (cause instanceof ConstraintViolationException cve) {
-      var constraintName = cve.getConstraintName();
-      var errorCode = getErrorCode(constraintName);
+      var errorCode = translate(cve);
       return buildResponseEntity(errorCode, VALIDATION_ERROR, UNPROCESSABLE_ENTITY);
     } else if (cause instanceof IllegalStateException ise) {
       var innerCause = ise.getCause();
