+++ src/main/java/org/folio/entlinks/utils/DatabaseConstraintTranslator.java
@@ -1,4 +1,4 @@
-package org.folio.entlinks.config.constants;
+package org.folio.entlinks.utils;
 
 import static org.folio.entlinks.config.constants.ErrorCode.DUPLICATE_AUTHORITY_ID;
 import static org.folio.entlinks.config.constants.ErrorCode.DUPLICATE_AUTHORITY_SOURCE_FILE_CODE;
@@ -7,15 +7,17 @@ import static org.folio.entlinks.config.constants.ErrorCode.DUPLICATE_AUTHORITY_
 import static org.folio.entlinks.config.constants.ErrorCode.DUPLICATE_AUTHORITY_SOURCE_FILE_SEQUENCE;
 import static org.folio.entlinks.config.constants.ErrorCode.DUPLICATE_AUTHORITY_SOURCE_FILE_URL;
 import static org.folio.entlinks.config.constants.ErrorCode.DUPLICATE_NOTE_TYPE_NAME;
-import static org.folio.entlinks.config.constants.ErrorCode.UNKNOWN_CONSTRAIN;
+import static org.folio.entlinks.config.constants.ErrorCode.UNKNOWN_CONSTRAINT;
 import static org.folio.entlinks.config.constants.ErrorCode.VIOLATION_OF_RELATION_BETWEEN_AUTHORITY_AND_SOURCE_FILE;
 import static org.folio.entlinks.config.constants.ErrorCode.VIOLATION_OF_RELATION_BETWEEN_AUTHORITY_ARCHIVE_AND_SOURCE_FILE;
 
 import java.util.Map;
+import org.folio.entlinks.config.constants.ErrorCode;
+import org.hibernate.exception.ConstraintViolationException;
 
-public class Constrains {
+public class DatabaseConstraintTranslator {
 
-  private static final Map<String, ErrorCode> DB_CONSTRAINS_I18N_MAP = Map.of(
+  private static final Map<String, ErrorCode> DB_CONSTRAINTS_I18N_MAP = Map.of(
     "authority_note_type_name_unq", DUPLICATE_NOTE_TYPE_NAME,
     "authority_source_file_name_unq", DUPLICATE_AUTHORITY_SOURCE_FILE_NAME,
     "authority_source_file_base_url_unq", DUPLICATE_AUTHORITY_SOURCE_FILE_URL,
@@ -27,7 +29,7 @@ public class Constrains {
     "pk_authority_source_file", DUPLICATE_AUTHORITY_SOURCE_FILE_ID
   );
 
-  public static ErrorCode getErrorCode(String constrainName) {
-    return DB_CONSTRAINS_I18N_MAP.getOrDefault(constrainName, UNKNOWN_CONSTRAIN);
+  public static ErrorCode translate(ConstraintViolationException cve) {
+    return DB_CONSTRAINTS_I18N_MAP.getOrDefault(cve.getConstraintName(), UNKNOWN_CONSTRAINT);
   }
 }
